+++ spec/dummy/app/models/custom_user.rb
@@ -1,0 +1,8 @@
+class CustomUser < User
+  validates :name, presence: true
+
+  def whodunnit
+    name = log_data.responsible_id
+    User.find_by!(name: name) if name.present?
+  end
+end