+++ spec/support/acceptance_helpers.rb
@@ -10,5 +10,10 @@ module Logidze
       expect(File.readlines(path).grep(/#{statement}/))
         .not_to be_empty, "File #{path} does not contain '#{statement}'"
     end
+
+    def verify_file_not_contain(path, statement)
+      expect(File.readlines(path).grep(/#{statement}/))
+        .to be_empty, "File #{path} should not contain '#{statement}'"
+    end
   end
 end