+++ spec/support/shared_contexts/cleanup_files_contexts.rb
@@ -1,7 +1,20 @@
 # frozen_string_literal: true
 shared_context "cleanup migrations" do
   before(:all) { @old_migrations = Dir["spec/dummy/db/migrate/*"] }
-  after(:all) { FileUtils.rm(Dir["spec/dummy/db/migrate/*"] - @old_migrations) }
+  after(:all) do
+    all_versions = ActiveRecord::Migrator.get_all_versions
+    (Dir["spec/dummy/db/migrate/*"] - @old_migrations).each do |path|
+      version = path.match(%r{\/(\d+)\_[^\.]+\.rb$})[1]
+      if all_versions.include?(version.to_i)
+        Dir.chdir("#{File.dirname(__FILE__)}/../../dummy") do
+          system <<-CMD
+            VERSION=#{version} rake db:migrate:down
+          CMD
+        end
+      end
+      FileUtils.rm(path)
+    end
+  end
 end
 
 shared_context "cleanup models" do