+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.3.0
+- Add `--update` option to install migration generator
+- Add `--only-trigger` option to model migration generator
+- Add [Responsibility](https://github.com/palkan/logidze/issues/4) feature
+
+# 0.2.3
+- Support Ruby >= 2.1
+
 # 0.2.2
 - Add `--backfill` option to model migration
 - Handle legacy data (that doesn't have log data)